chore(deps): upgrade hcloud-go to v2.44.0
Bump github.com/hetznercloud/hcloud-go from v1.33.0 to v2.44.0 (new module
path v2/hcloud). This unblocks the type-availability-gate plan, which
depends on v2's ServerType.Locations[]ServerTypeLocation{Name, Available}.

v1 -> v2 changes that needed code adjustments:

1. GetByID signatures on ServerClient and ActionClient now take int64
   instead of int (matching the new int64 Server.ID / Action.ID fields).
2. Server.ID, Action.ID, ServerType.ID, Datacenter.ID, Location.ID,
   Image.ID etc. widened from int to int64.
3. The Go directive in go.mod bumped from 1.23 to 1.25.0 (required by
   hcloud-go v2.44.0, which depends on golang.org/x 1.25-line packages).

To keep the engine decoupled from the SDK's exact ID type, the
internal/hetzner.API interface methods (GetServer, GetAction) were
widened from int to int64 too. This is a behavior-neutral type
widening: the wrapper now matches the SDK's ID type at its boundary,
and call sites that hold IDs in store.Server.HCloudServerID (int)
cast with int64(...) where they cross into the wrapper. Internal
maps (hcloudmock.Fake.servers / actions) and their key types followed.

No behavior changes in the rescaler, scheduler, or API. All existing
tests pass; go vet is clean.

+++ cmd/try.go
@@ -71,7 +71,7 @@ func runTry(cmd *cobra.Command, args []string) {
 		return
 	}
 
-	hsrv, err := api.GetServer(context.Background(), srv.HCloudServerID)
+	hsrv, err := api.GetServer(context.Background(), int64(srv.HCloudServerID))
 	if err != nil {
 		fmt.Println("get server:", err)
 		return
